TASK-178: fix bug - show  deposit amount in USD, move policy to p2p review if smart deposit received via smart deposit status  changing
