+++ dynamis/apps/accounts/templates/accounts/smart-deposit-stub.html
@@ -36,8 +36,8 @@
                     <div class="input-field " type="hidden">
                         <div class="input-field col s12" type="hidden">
                             <input class=" validate" id="id_amount" name="amount"
-                                   value="{{ user.policies.last.smart_deposit.cost }}" step="any" type="number"/>
-                            <label class="s12 " for="id_amount">Amount to send, ETH</label>
+                                   value="{{ user.policies.last.smart_deposit.cost_dollar }}" step="any" type="number"/>
+                            <label class="s12 " for="id_amount">Amount to send, USD</label>
                         </div>
                     </div>
                     <div class="input-field " type="hidden">
