+++ tests/python/policies/conftest.py
@@ -2,24 +2,6 @@ import pytest
 
 
 @pytest.fixture
-def policy_data():
-    policy_data = {
-        'identity': {
-            "verification_method": "keybase",
-            "verification_data": {
-                "username": "test",
-                "proofs": [],
-            },
-        },
-        'employmentHistory': {
-            'jobs': [],
-        },
-        'questions': {}
-    }
-    return policy_data
-
-
-@pytest.fixture
 def job_data():
     job_data = {
         'company': "test company",
